'runTheSPEACWeightings' works!
Two functions, 'getRootMotionWeightings' and 'getChordRoots' behave
differently than expected in the documentation examples - the latter
function is most certainly responsible for the misbehavior of the
former.
Yet 'getRootmotionWeightings' does behave as expected in the book so
this is probably not a bug.

speac.lisp contains a comment on this in lines 609 to 614.